requirements/prod.txt
# Production module requirements go here.
# Be sure that you always specify an exact version such as:
# foo==1.2.3 and you list all dependencies explicitly.

-r compiled.txt

boto==2.23.0
cef==0.5
commonware==0.4.2
Django==1.6.1
django-filter==0.7
django-paranoia==0.1.8.7
django-statsd-mozilla==0.3.10
djangorestframework==2.3.12
mozilla-logger==0.2
raven==4.0.3
requests==2.2.1
statsd==2.1.2

# Waiting on a 0.1.4 PyPI release.
-e git://github.com/mozilla/PyHawk.git@91bb041806499a9ecb5bc2bf9f2bbecf5976caa4#egg=PyHawk
